+++ docs/components/Flow/demos/component_options.py
@@ -2,8 +2,10 @@ import json
 import os
 
 import gradio as gr
+
 import modelscope_studio as mgr
-from modelscope_studio.components.Flow import BackgroundPropsDict, Edge, Node
+from modelscope_studio.components.legacy.Flow import (BackgroundPropsDict,
+                                                      Edge, Node)
 
 with open((os.path.join(os.path.dirname(__file__),
                         "../schema/agents_schema.json"))) as f: