+++ frontend/legacy/Flow/shared/utils.ts
@@ -1,4 +1,4 @@
-import type { FlowRenderData } from '@modelscope-studio/compiled';
+import type { FlowRenderData } from '@modelscope-studio/legacy-compiled';
 
 export interface FlowData {
   nodes?: NonNullable<FlowRenderData['nodes']>;