+++ frontend/legacy/MultimodalInput/shared/FilePreview.svelte
@@ -2,7 +2,7 @@
   import IconButton from '@gradio/atoms/src/IconButton.svelte';
   import type { FileData } from '@gradio/client';
   import type { I18nFormatter, SelectData } from '@gradio/utils';
-  import { FileView } from '@modelscope-studio/compiled';
+  import { FileView } from '@modelscope-studio/legacy-compiled';
   import { createEventDispatcher } from 'svelte';
 
   import Loader from './Loader.svelte';